feat: add new improved version
